requirements.txt
fastapi
pydantic
pymongo
python-dotenv
openai
uvicorn
pymongo[srv]
certifi
